lesson 20_1. ДЗ END Добавил double при делении.

+++ src/main/java/by/valery/springcourseMVC/controllers/FistController.java
@@ -31,7 +31,7 @@ public class FistController {
            model.addAttribute("result", a - b);
        }
        else if(action.equals("division")) {
-           model.addAttribute("result", a / b);
+           model.addAttribute("result", (double)a / b);
        }
 
         return ("first/calculator");
